+++ flexmeasures/data/models/generic_assets.py
@@ -1142,6 +1142,8 @@ class GenericAsset(db.Model, AuthModelMixin):
         as_json: bool = False,
         compress_json: bool = False,
         resolution: timedelta | None = None,
+        use_materialized_view: bool = True,
+        include_live_tail: bool | None = None,
     ) -> BeliefsDataFrame | str:
         """Search all beliefs about events for all sensors of this asset
 
@@ -1186,6 +1188,8 @@ class GenericAsset(db.Model, AuthModelMixin):
                 most_recent_events_only=most_recent_events_only,
                 one_deterministic_belief_per_event_per_source=True,
                 resolution=resolution,
+                use_materialized_view=use_materialized_view,
+                include_live_tail=include_live_tail,
             )
         if as_json and not compress_json:
             from flexmeasures.data.services.time_series import simplify_index